Create Orthomcl analysis
Orthology-based analysis of the species’ gene complements to identify abundant species- and lineage-specific adaptations. To this end, reciprocal BLAST analysis (e-value: 1e-5, alignment coverage > 50%) as implemented in OrthoMCL was used to retrieve the set of orthologous protein-coding gene groups among the compared fungal genomes to determine the core, lineage- and species-specific gene families and genes. Here, “core” is defined as the gene families that are shared by all species and “lineage-specific” as the subset of core gene families that are not present in any other fungus. We considered “species-specific” as genes that are found in only one of the  species (among the compared genomes), while “orphans” are the subcategory of species-specific genes that do not have homologs in the other fungal species.